refactor: share blob download helper across settings and backup pages
